pyproject.toml
[tool.poetry]
name = "suggestions bot"
version = "3.28"
description = "A Discord bot designed to build better communities by encouraging a positive and constructive relationship between community and staff."
authors = ["Suggestions Bot"]
license = "AGPL-3.0"
readme = "readme.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
disnake = {extras = ["speed"], version = "2.10.1"}
alaric = "^1.5.0"
function-cooldowns = "^2.0.1"
httpx = "^0.28.1"
orjson = "^3.10.11"
skelmis-commons = "^1.4.0"
zonis = "^2.1.0"
aiobotocore = "^2.15.2"
logoo = "^1.5.1"
typing-extensions = "^4.12.2"
python-dotenv = "^1.0.1"
humanize = "^4.11.0"
disnake-ext-components = {git = "https://github.com/DisnakeCommunity/disnake-ext-components.git", rev = "rewrite"}


[tool.poetry.group.dev.dependencies]
causar = "^0.2.0"
mongomock-motor = "^0.0.34"
mongomock = "^4.2.0.post1"
motor-stubs = "^1.7.1"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
